requirements.txt
django-cors-headers
djangorestframework
djangorestframework-gis
drf-extensions
django-filter
pytz
mysqlclient